requirements.txt
youtube-transcript-api==0.6.2
tiktoken==0.7.0
langchain_openai==0.1.17
python-dotenv==1.0.1
regex==2024.5.15
pytubefix==6.13.0